fix notebook
