+++ experiments/demo_logs.jsonl
@@ -1,0 +1,1 @@
+{"experiment_id": "exp", "round": 0, "system": "system", "agent_id": "generator", "agent_role": "AgentRole.GENERATOR", "agent_engine": null, "symbol": "tmp_demo.py", "prompt": "format code", "response": null, "attempt_number": 0, "agent_action_outcome": "success", "start": "2025-05-18T11:23:41.641821+00:00", "stop": "2025-05-18T11:23:41.927913+00:00"}
