+++ tmp_demo.py
@@ -1,0 +1,1 @@
+x = 1
